requirements.txt
azureml-sdk[explain,automl,contrib,notebooks,tensorboard]
tensorflow
keras
transformers
matplotlib
pandas_profiling